+++ src/types/index.ts
@@ -52,3 +52,35 @@ export type PrettifyObject<T> = {
  * AtLeastOne<{ create?: string; update?: string; }> => ensures we have 'create' and/or 'update' defined
  */
 export type AtLeastOne<T, U = { [K in keyof T]: Pick<T, K> }> = Partial<T> & U[keyof U];
+
+// ======================== TEST TYPES =========================== //
+
+export type Expect<T extends true> = T;
+export type ExpectTrue<T extends true> = T;
+export type ExpectFalse<T extends false> = T;
+export type IsTrue<T extends true> = T;
+export type IsFalse<T extends false> = T;
+
+export type Equal<X, Y> = (<T>() => T extends X ? 1 : 2) extends <T>() => T extends Y ? 1 : 2
+  ? true
+  : false;
+export type NotEqual<X, Y> = true extends Equal<X, Y> ? false : true;
+
+export type NotAny<T> = true extends IsAny<T> ? false : true;
+
+export type Debug<T> = { [K in keyof T]: T[K] };
+export type MergeInsertions<T> = T extends object ? { [K in keyof T]: MergeInsertions<T[K]> } : T;
+
+export type Alike<X, Y> = Equal<MergeInsertions<X>, MergeInsertions<Y>>;
+
+export type ExpectExtends<VALUE, EXPECTED> = EXPECTED extends VALUE ? true : false;
+export type ExpectValidArgs<
+  FUNC extends (...args: any[]) => any,
+  ARGS extends any[],
+> = ARGS extends Parameters<FUNC> ? true : false;
+
+export type UnionToIntersection<U> = (U extends any ? (k: U) => void : never) extends (
+  k: infer I,
+) => void
+  ? I
+  : never;